+++ FlightManager/Web/Models/Flights/FlightAdminListViewModel.cs
@@ -19,7 +19,7 @@ namespace Web.Models.Users
 
         public string PlaneType { get; set; }
 
-        public int PlaneNumber { get; set; }
+        public long PlaneNumber { get; set; }
 
         public string PilotName { get; set; }
